requirements.txt
fastapi
uvicorn[standard]
alembic
SQLAlchemy
python-jose
passlib[bcrypt]
pyjwt
python-socketio
redis
websockets